refactor: organize files in `hyperlane-ethereum`
