+++ rust/chains/hyperlane-ethereum/.gitignore
@@ -1,1 +1,1 @@
-src/contracts
+src/interfaces
