+++ rust/chains/hyperlane-ethereum/build.rs
@@ -1,3 +1,3 @@
 fn main() {
-    abigen::generate_bindings_for_dir("./abis", "./src/contracts", abigen::BuildType::Ethers);
+    abigen::generate_bindings_for_dir("./abis", "./src/interfaces", abigen::BuildType::Ethers);
 }
