+++ rust/chains/hyperlane-ethereum/src/contracts/mod.rs
@@ -1,0 +1,9 @@
+pub use {interchain_gas::*, mailbox::*, merkle_tree_hook::*, validator_announce::*};
+
+mod interchain_gas;
+
+mod mailbox;
+
+mod merkle_tree_hook;
+
+mod validator_announce;
