+++ rust/chains/hyperlane-ethereum/src/lib.rs
@@ -9,70 +9,25 @@ use ethers::abi::FunctionExt;
 use ethers::prelude::{abi, Lazy, Middleware};
 
 #[cfg(not(doctest))]
-pub use self::{
-    aggregation_ism::*, ccip_read_ism::*, config::*, config::*, interchain_gas::*,
-    interchain_gas::*, interchain_security_module::*, interchain_security_module::*, mailbox::*,
-    mailbox::*, merkle_tree_hook::*, multisig_ism::*, provider::*, routing_ism::*, rpc_clients::*,
-    signers::*, singleton_signer::*, trait_builder::*, validator_announce::*,
-};
+pub use self::{config::*, contracts::*, ism::*, rpc_clients::*, signer::*};
 
 #[cfg(not(doctest))]
 mod tx;
 
-/// Mailbox abi
 #[cfg(not(doctest))]
-mod mailbox;
+mod contracts;
 
 #[cfg(not(doctest))]
-mod trait_builder;
-
-/// Provider abi
-#[cfg(not(doctest))]
-mod provider;
-
-/// InterchainGasPaymaster abi
-#[cfg(not(doctest))]
-mod interchain_gas;
-
-/// interchain_security_module abi
-#[cfg(not(doctest))]
-mod interchain_security_module;
-
-/// Merkle tree hook abi
-#[cfg(not(doctest))]
-mod merkle_tree_hook;
-
-/// MultisigIsm abi
-#[cfg(not(doctest))]
-mod multisig_ism;
-
-/// RoutingIsm abi
-#[cfg(not(doctest))]
-mod routing_ism;
-
-/// CcipReadIsm abi
-#[cfg(not(doctest))]
-mod ccip_read_ism;
-
-/// ValidatorAnnounce abi
-#[cfg(not(doctest))]
-mod validator_announce;
-
-/// AggregationIsm abi
-#[cfg(not(doctest))]
-mod aggregation_ism;
+mod ism;
 
 /// Generated contract bindings.
 #[cfg(not(doctest))]
-mod contracts;
+mod interfaces;
 
 /// Ethers JSONRPC Client implementations
 mod rpc_clients;
 
-mod signers;
-
-#[cfg(not(doctest))]
-mod singleton_signer;
+mod signer;
 
 mod config;
 mod error;
