+++ rust/chains/hyperlane-ethereum/src/signer/singleton.rs
